requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Bot Framework
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0

# Web Scraping (optional)
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0

# Email Monitoring (optional)
aiosmtplib>=3.0.0
aioimaplib>=1.0.0
